requirements.txt
streamlit
requests
newspaper3k
python-dotenv
google-generativeai
